Reorganised facet builder code - moved enum inside builder class and removed the jsonfacets package.
